Fixed non-compiling code & minor warning
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@4948 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/client/CWeather.cpp
@@ -107,7 +107,7 @@ void CWeather::SimulateSnow(float dt)
     m_fWind += dt*5;
     if( m_fWind > 360 )
         m_fWind -= 360;
-    float wind = sinf( m_fWind * PI/180 );
+    float wind = sinf( m_fWind * (float)PI/180 );
 
 
     // Simulate the particles
